src/mcp-agent/requirements.txt
httpx>=0.27.0
pyjwt>=2.8.0
cryptography>=42.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
spiffe>=0.2.3
